+++ packages/kysely/migrate/migrations/2023-11-07T23:18:12_passwords.ts
@@ -4,14 +4,9 @@ import { Database } from "../../src/db";
 export async function up(db: Kysely<Database>): Promise<void> {
   await db.schema
     .createTable("passwords")
-    // .addColumn("tenant_id", "varchar(255)", (col) =>
-    //   col.references("tenants.id").onDelete("cascade").notNull(),
-    // )
-    // .addColumn("user_id", "varchar(255)", (col) =>
-    //   col.references("user.id").onDelete("cascade").primaryKey(),
-    // )
     .addColumn("tenant_id", "varchar(255)")
     .addColumn("user_id", "varchar(255)")
+    .addPrimaryKeyConstraint("passwords_pkey", ["user_id", "tenant_id"])
     .addForeignKeyConstraint(
       "user_id_constraint",
       ["user_id", "tenant_id"],
@@ -26,15 +21,8 @@ export async function up(db: Kysely<Database>): Promise<void> {
   await db.schema
     .createTable("codes")
     .addColumn("id", "varchar(255)", (col) => col.primaryKey())
-    // .addColumn("tenant_id", "varchar(255)", (col) =>
-    //   col.references("tenants.id").onDelete("cascade").notNull(),
-    // )
     .addColumn("user_id", "varchar(255)")
     .addColumn("tenant_id", "varchar(255)")
-    // fk mismatch - we were referencing user - singular - non-existent table
-    // .addColumn("user_id", "varchar(255)", (col) =>
-    //   col.references("user.id").onDelete("cascade").notNull(),
-    // )
     .addForeignKeyConstraint(
       "user_id_constraint",
       ["user_id", "tenant_id"],