+++ packages/kysely/migrate/migrations/2024-07-26T10:18:00_codes.ts
@@ -23,7 +23,7 @@ export async function up(db: Kysely<Database>): Promise<void> {
     .addColumn("updated_at", "varchar(255)", (col) => col.notNull())
     .addColumn("expires_at", "varchar(255)", (col) => col.notNull())
     .addColumn("ip", "varchar(255)")
-    .addColumn("useragent", "varchar(255)")
+    .addColumn("useragent", "varchar(512)")
     .execute();
 
   await db.schema