+++ packages/kysely/migrate/migrations/index.ts
@@ -66,6 +66,7 @@ import * as n65_code_verifier from "./2024-11-22T17:48:00_code_verifier";
 import * as n66_email_providers from "./2024-12-04T23:34:00_email_provider";
 import * as n67_drop_tickets from "./2024-12-05T13:05:00_drop_tickets";
 import * as n68_login_useragents from "./2024-12-05T13:20:00_login_useragent";
+import * as n70_refresh_tokens from "./2025-02-03T08:48:00_refresh_tokens";
 
 // These need to be in alphabetic order
 export default {
@@ -137,4 +138,5 @@ export default {
   n66_email_providers,
   n67_drop_tickets,
   n68_login_useragents,
+  n70_refresh_tokens,
 };