+++ packages/adapter-interfaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @authhero/adapter-interfaces
 
+## 0.38.0
+
+### Minor Changes
+
+- Add table for refresh tokens
+
 ## 0.37.0
 
 ### Minor Changes