+++ packages/drizzle/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @authhero/drizzle
 
+## 0.1.75
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.38.0
+
 ## 0.1.74
 
 ### Patch Changes